restructering the database files
