+++ src/controllers/event.js
@@ -1,9 +1,9 @@
-const events = require('./../../database/db_connection.js');
-
-exports.get = (req, res, next) => {
-  const { event } = req.params;
-  if (events.includes(event)) {
-    return res.render('event', { event });
-  }
-  next();
-};
+// const events = require('./../../database/db_connection.js');
+//
+// exports.get = (req, res, next) => {
+//   const { event } = req.params;
+//   if (events.includes(event)) {
+//     return res.render('event', { event });
+//   }
+//   next();
+// };
