+++ src/server.js
@@ -1,5 +1,8 @@
 const app = require('./app.js');
-
-app.listen(app.get('port'), () => {
-  console.log('Magic happens on port 5000!');
+const db = require('./database/db_connection.js');
+db.once('open', function () {
+  console.log('connected to DB');
+  app.listen(app.get('port'), () => {
+    console.log('Magic happens on port 5000!');
+  });
 });
